Allow Arduino code to use Serial0
For now Arduino code can *not* *read* from the serial.
We need to find out how to disable the Contiki receiving side and enable
the Arduino side instead. We can't implement the Arduino API in terms of
the Contiki API since the latter is not powerful enough.

Still has some rough edges in that the availability for sending is
always 1 byte. So if some application out there waits for more bytes for
sending it will wait forever.